Move frontend layer to `frontend` directory
