+++ .vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "ee67688b",
+  "configHash": "6cb166c5",
+  "lockfileHash": "e3b0c442",
+  "browserHash": "deca6459",
+  "optimized": {},
+  "chunks": {}
+}
