+++ frontend/src/components/layout/NavBar.tsx
