+++ frontend/src/pages/Browse/FiltersPanel.tsx
