+++ frontend/src/pages/Compare/Compare.css
