+++ frontend/src/pages/Compare/Compare.tsx
